requirements.txt
openai>=1.0.0
pyautogui>=0.9.54
pyyaml>=6.0
pyperclip>=1.8.0
pillow>=10.0.0
